move the main cluster control CPUs to preemptible
We then have to move the utility pods to deployment, to ensure
they are rescheduled as needed. Note that we should have done
that in the first place, so it is also a win.

The price is 3x lower, and we don't really care (once the
connection between the master and slave is created, we don't
rely on the control pods)