export extra info in both the VM/container image and the report
This adds a UUID to the VM/container image and export it to the build
report. It will allow to match the two after the fact for backends
that wouldn't allow us to store and access it in another way.

Moreover, it exports the versions (with fingerprint) in the build
report, which will probably be very useful for debugging/reproduceability