fix access scheme for the import drive to allow sharing
This change comes straight from the K8s/GKE documentation. Seems
to be the only way to have a drive mounted in multiple pods at the
same time. The trick seems to be that the GKE volume must be mounted
read-only, or GKE won't allow mounting it in multiple VMs.

In addition, import cache locks now need to be builder-global.
The contention happens on K8s itself, with GKE refusing to mount
the path multiple times (i.e. container init fails).